[0c901cb] settings: move settings library files into separate folder

+++ xbmc/settings/Settings.cpp
@@ -47,7 +47,7 @@
 #include "settings/MediaSourceSettings.h"
 #include "settings/SettingAddon.h"
 #include "settings/SettingControl.h"
-#include "settings/SettingsManager.h"
+#include "settings/lib/SettingsManager.h"
 #include "settings/SettingPath.h"
 #include "settings/SkinSettings.h"
 #include "threads/SingleLock.h"
@@ -547,7 +547,7 @@ bool CSettings::SetList(const std::string &id, const std::vector<CVariant> &valu
       case SettingTypeInteger:
         if (!itValue->isInteger())
           return false;
-        ret = static_cast<CSettingInt*>(settingValue)->SetValue(itValue->asInteger());
+        ret = static_cast<CSettingInt*>(settingValue)->SetValue((int)itValue->asInteger());
         break;
 
       case SettingTypeNumber:
